# Fix slow qualified name perf in 0.4.2+