--- a/libcst/metadata/tests/test_scope_provider.py
+++ b/libcst/metadata/tests/test_scope_provider.py
@@ -1445,35 +1445,33 @@ class ScopeProviderTest(UnitTest):
 
     def test_global_contains_is_read_only(self) -> None:
         gscope = GlobalScope()
-        before_assignments = list(gscope._assignments.items())
-        before_accesses = list(gscope._accesses.items())
+        before_assignments = list(gscope.assignments)
+        before_accesses = list(gscope.accesses)
         self.assertFalse("doesnt_exist" in gscope)
-        self.assertEqual(list(gscope._accesses.items()), before_accesses)
-        self.assertEqual(list(gscope._assignments.items()), before_assignments)
+        self.assertEqual(list(gscope.accesses), before_accesses)
+        self.assertEqual(list(gscope.assignments), before_assignments)
 
     def test_contains_is_read_only(self) -> None:
         for s in [LocalScope, FunctionScope, ClassScope, ComprehensionScope]:
             with self.subTest(scope=s):
                 gscope = GlobalScope()
                 scope = s(parent=gscope, node=cst.Name("lol"))
-                before_assignments = list(scope._assignments.items())
-                before_accesses = list(scope._accesses.items())
+                before_assignments = list(scope.assignments)
+                before_accesses = list(scope.accesses)
                 before_overwrites = list(scope._scope_overwrites.items())
-                before_parent_assignments = list(scope.parent._assignments.items())
-                before_parent_accesses = list(scope.parent._accesses.items())
+                before_parent_assignments = list(scope.parent.assignments)
+                before_parent_accesses = list(scope.parent.accesses)
 
                 self.assertFalse("doesnt_exist" in scope)
-                self.assertEqual(list(scope._accesses.items()), before_accesses)
-                self.assertEqual(list(scope._assignments.items()), before_assignments)
+                self.assertEqual(list(scope.accesses), before_accesses)
+                self.assertEqual(list(scope.assignments), before_assignments)
                 self.assertEqual(
                     list(scope._scope_overwrites.items()), before_overwrites
                 )
                 self.assertEqual(
-                    list(scope.parent._assignments.items()), before_parent_assignments
+                    list(scope.parent.assignments), before_parent_assignments
                 )
-                self.assertEqual(
-                    list(scope.parent._accesses.items()), before_parent_accesses
-                )
+                self.assertEqual(list(scope.parent.accesses), before_parent_accesses)
 
     def test_attribute_of_function_call(self) -> None:
         get_scope_metadata_provider("foo().bar")
@@ -1496,11 +1494,11 @@ class ScopeProviderTest(UnitTest):
         )
         a = m.body[0]
         scope = scopes[a]
-        assignments_len_before = len(scope._assignments)
-        accesses_len_before = len(scope._accesses)
+        assignments_before = list(scope.assignments)
+        accesses_before = list(scope.accesses)
         scope.get_qualified_names_for("doesnt_exist")
-        self.assertEqual(len(scope._assignments), assignments_len_before)
-        self.assertEqual(len(scope._accesses), accesses_len_before)
+        self.assertEqual(list(scope.assignments), assignments_before)
+        self.assertEqual(list(scope.accesses), accesses_before)
 
     def test_gen_dotted_names(self) -> None:
         names = {name for name, node in _gen_dotted_names(cst.Name(value="a"))}