## Summary
This is a performance fix for a change I made to correct compute qualified names in #682

## Test Plan

I hacked three versions of the `node_accesses` set creation in `Scope.get_qualified_names_for` and ran a perf test on them:

```
(libcstvenv) lpetre@lpetre-mbp LibCST % cat /tmp/fastero.json 
{
    "setup": "from libcst import safe_parse_module, MetadataWrapper; from libcst.metadata import QualifiedNameProvider; import os; from wcmatch import wcmatch; env = os.environ; paths = list(wcmatch.WcMatch('/Users/lpetre/dev/src/github.com/python/cpython/Lib', '*.py', exclude_pattern='tests|test', flags=wcmatch.RECURSIVE).imatch()); paths = paths[:100]; mods = [MetadataWrapper(safe_parse_module(open(path).read())) for path in paths];",
    "results": [
        {
            "snippet_name": "0.4.2",
            "snippet_code": "env['FIX_QNAME'] = '0.4.2'; [mod.resolve(QualifiedNameProvider) for mod in mods]"
         },
         {
            "snippet_name": "0.4.4",
            "snippet_code": "env['FIX_QNAME'] = '0.4.4'; [mod.resolve(QualifiedNameProvider) for mod in mods]"
         },
         {
            "snippet_name": "0.4.5",
            "snippet_code": "env['FIX_QNAME'] = '0.4.5'; [mod.resolve(QualifiedNameProvider) for mod in mods]"
         }
    ]
 }

(libcstvenv) lpetre@lpetre-mbp LibCST % python -m fastero -f /tmp/fastero.json -m 5
╭───────────────────────────── Setup code ─────────────────────────────╮
│   1 from libcst import safe_parse_module, MetadataWrapper; from      │
│     libcst.metadata import QualifiedNameProvider; import os; from    │
│     wcmatch import wcmatch; env = os.environ; paths = list(wcmatch.W │
│     Match('/Users/lpetre/dev/src/github.com/python/cpython/Lib',     │
│     '*.py', exclude_pattern='tests|test',                            │
│     flags=wcmatch.RECURSIVE).imatch()); paths = paths[:100]; mods =  │
│     [MetadataWrapper(safe_parse_module(open(path).read())) for path  │
│     in paths];                                                       │
╰──────────────────────────────────────────────────────────────────────╯
────────────────────────────────────────────────────────────────────────────────────────────────────────────── Benchmark started… ──────────────────────────────────────────────────────────────────────────────────────────────────────────────
0.4.2: env['FIX_QNAME'] = '0.4.2'; [mod.resolve(QualifiedNameProvider) for mod in mods]
  Time  (mean ± σ):       34.785 s ± 19.85 ms
  Range (min  … max):     34.753 s … 34.805 s    [runs: 5]
0.4.4: env['FIX_QNAME'] = '0.4.4'; [mod.resolve(QualifiedNameProvider) for mod in mods]
  Time  (mean ± σ):        38.538 s ± 628.35 ms
  Range (min  … max):      37.862 s …  39.197 s    [runs: 5]
0.4.5: env['FIX_QNAME'] = '0.4.5'; [mod.resolve(QualifiedNameProvider) for mod in mods]
  Time  (mean ± σ):       34.278 s ± 50.14 ms
  Range (min  … max):     34.224 s … 34.339 s    [runs: 5]

Summary:
┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ Bar Chart ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
┃ 0.4.2 [34.753 s]: ▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆     ┃
┃ 0.4.4 [37.862 s]: ▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆ ┃
┃ 0.4.5 [34.224 s]: ▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆▆      ┃
┗━━━━━━━━━━━━━━━━━━━━━━━━━ (lower is better) ━━━━━━━━━━━━━━━━━━━━━━━━━━┛
  0.4.5 is the fastest.
    1.01 (1.02 … 1.01) times faster than 0.4.2
    1.12 (1.11 … 1.14) times faster than 0.4.4
```